move thread_specific_ptr into verus_clhash.h
