+++ src/crypto/verus_clhash.h
@@ -70,7 +70,29 @@ struct verusclhash_descr
     uint32_t keySizeInBytes;
 };
 
-struct thread_specific_ptr;
+struct thread_specific_ptr {
+    void *ptr;
+    thread_specific_ptr() { ptr = nullptr; }
+    void reset(void *newptr = nullptr)
+    {
+        if (ptr && ptr != newptr)
+        {
+            std::free(ptr);
+        }
+        ptr = newptr;
+
+    }
+    void *get() { return ptr; }
+#if defined(__APPLE__) || defined(_WIN32)
+    // horrible MingW and Mac with gcc thread local storage bug workaround
+    ~thread_specific_ptr();
+#else
+    ~thread_specific_ptr() {
+        this->reset();
+    }
+#endif
+};
+
 extern thread_local thread_specific_ptr verusclhasher_key;
 extern thread_local thread_specific_ptr verusclhasher_descr;
